IOS-5745 Fix aptos review comments
