--- BlockchainSdk/BlockchainService/Aptos/DTO/AptosResponse+Account.swift
@@ -1,11 +1,0 @@
-//
-//  AptosResponse+Account.swift
-//  BlockchainSdk
-//
-//  Created by skibinalexander on 26.01.2024.
-//  Copyright © 2024 Tangem AG. All rights reserved.
-//
-
-import Foundation
-
-extension AptosResponse {}
