+++ BlockchainSdk/BlockchainService/Aptos/DTO/AptosResponse.swift
@@ -9,3 +9,22 @@
 import Foundation
 
 enum AptosResponse {}
+
+/// https://aptos.dev/concepts/resources#resources-vs-instances
+extension AptosResponse {
+    struct CoinStore: Decodable {
+        let coin: Coin
+    }
+    
+    struct Coin: Decodable {
+        let value: UInt64
+    }
+}
+
+extension AptosResponse {
+    struct Fee: Decodable {
+        let deprioritizedGasEstimate: UInt64
+        let gasEstimate: UInt64
+        let prioritizedGasEstimate: UInt64
+    }
+}
